refactor(telephony): move _CallState + orchestration helpers to session.py (#251)
Slims router.py to FastAPI shells + WS event-loop dispatch
(~613 lines). All call orchestration — _CallState, barge-in, hangup,
silence, LLM turn, transcript consumer, chunking + goodbye-detection
helpers, _bg_call_event — now lives in session.py (~656 lines).

Test imports updated to the new module path. No behavior change.
